pyproject.toml
[tool.poetry]
name = "govready_cub"
version = "0.1.0"
description = "GovReady CaC Utility Belt"
authors = ["Greg Elin <greg.elin@govready.com>", "Tom Wood <tom.wood@civicactions.com>"]
license = "GPL-3.0-or-later"

[tool.poetry.dependencies]
python = "^3.8"
click = "^7.1.2"
simplejson = "^3.17.2"
spacy = "^2.3.4"
python-slugify = "^4.0.1"
pydantic = "^1.7.3"
jsonschema = "^3.2.0"
graph-theory = "^2020.12.17"

[tool.poetry.dev-dependencies]
pytest = "^6.1.2"
pip-tools = "^5.4.0"
mypy = "^0.790"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
